[DEPYLER-0276] Fix CSE Type Preservation for len() Calls
Fixed CSE optimization breaking when extracting len() calls that need type conversion.

Root Cause:
- DEPYLER-0275 removed cast from convert_len_call() to prevent double casts
- But CSE optimization runs BEFORE return statement processing
- When CSE extracts len(), it creates usize temp variable without cast
- Target variable expects i32, causing type mismatch compilation error

Fix Applied:
- Reverted cast removal in expr_gen.rs:690-701
- Keep "as i32" cast in convert_len_call() for CSE compatibility
- Accept that some return statements might have double casts
- CSE compatibility is more important than avoiding occasional double casts

Changes:
- crates/depyler-core/src/rust_gen/expr_gen.rs: Restored "as i32" cast
- docs/bugs/DEPYLER-0276-CSE-TYPE-PRESERVATION.md: Comprehensive bug analysis
- examples/showcase/binary_search.rs: Re-transpiled with fix (now compiles)
- CHANGELOG.md: Added v3.19.1 entry documenting fix

Verification:
- binary_search.py: transpiles and compiles with zero errors/warnings
- contracts_example.py: transpiles and compiles with zero errors/warnings
- Generated code shows: let _cse_temp_0 = arr.len() as i32;

Extreme TDD Cycle:
- RED: binary_search.rs failed to compile (expected i32, found usize)
- GREEN: Restored cast in convert_len_call() → compiles successfully
- REFACTOR: Validated on showcase examples, all pass

+++ crates/depyler-core/src/rust_gen/expr_gen.rs
@@ -693,11 +693,12 @@ impl<'a, 'b> ExpressionConverter<'a, 'b> {
         }
         let arg = &args[0];
 
-        // DEPYLER-0275: Return .len() without cast - let caller add cast if needed
-        // Python's len() → Rust's .len() which returns usize
-        // The return statement processing in stmt_gen.rs will add `as i32` if needed
-        // This prevents double casts like `s.len() as i32 as i32`
-        Ok(parse_quote! { #arg.len() })
+        // DEPYLER-0276: Keep cast for CSE compatibility
+        // Python's len() returns int (maps to i32)
+        // Rust's .len() returns usize, so we cast to i32
+        // CSE optimization runs before return statement processing, so we need the cast here
+        // to avoid type mismatches when CSE extracts len() into a temporary variable
+        Ok(parse_quote! { #arg.len() as i32 })
     }
 
     fn convert_int_cast(&self, hir_args: &[HirExpr], arg_exprs: &[syn::Expr]) -> Result<syn::Expr> {
